cleanup javascript code
