+++ app/views/plans/new.html.erb
@@ -3,7 +3,7 @@
   <!-- component -->
   <div class="flex flex-col items-center justify-center p-12 mx-auto" data-controller="request">
     <div class="w-full">
-      <%= form_with model: @plan, data: {controller: "send"} do |form| %>
+      <%= form_with model: @plan, data: {controller: "edit"} do |form| %>
         <div class="mb-5">
           <%= form.label :name, "行程名稱", class: "mb-3 block text-base font-medium text-blue-600" %>
           <%= form.text_field :name, class: "name w-full rounded-md border border-[#e0e0e0] bg-white py-3 px-6 text-base font-medium text-[#6B7280] outline-none focus:border-[#6A64F1] focus:shadow-md"%>
@@ -46,8 +46,8 @@
             <% end %>
           </div>
         <% end %>
-        <div data-action="click->send#update" class="inline-block mx-auto my-10 shadow bg-indigo-600 hover:bg-indigo-700 focus:shadow-outline focus:outline-none text-white text-xs py-3 px-10 rounded">更新行程</div>
-      </div>
-    <% end %>
+        <button data-action="click->edit#update" class="inline-block mx-auto my-10 shadow bg-indigo-600 hover:bg-indigo-700 focus:shadow-outline focus:outline-none text-white text-xs py-3 px-10 rounded">更新行程</button>
+      <% end %>
+    </div>
   </div>
 </div>
